Added allowed domains FAQ
Jim, contact me if you have questions about this. We got a ticket about token "allowed domains", and then Ben explained it to us in the tech-support channel. It was then suggested we make a FAQ out of it. This customer (and me, really) didn't realize "allowed domains" has to match a header called "Referer" in the request.